requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.9
nibabel>=5.2
scikit-image>=0.22
matplotlib>=3.7
pyyaml>=6.0
joblib>=1.2
xgboost>=1.7
